+++ PrintFarmer/Navigation/AppRouter.swift
@@ -7,6 +7,7 @@ final class AppRouter {
     var printersPath = NavigationPath()
     var jobsPath = NavigationPath()
     var notificationsPath = NavigationPath()
+    var inventoryPath = NavigationPath()
     var notificationBadgeCount: Int = 0
 
     func resetToRoot(tab: AppTab) {
@@ -15,6 +16,7 @@ final class AppRouter {
         case .printers: printersPath = NavigationPath()
         case .jobs: jobsPath = NavigationPath()
         case .notifications: notificationsPath = NavigationPath()
+        case .inventory: inventoryPath = NavigationPath()
         case .settings: break
         }
     }
@@ -25,5 +27,6 @@ enum AppTab: Hashable {
     case printers
     case jobs
     case notifications
+    case inventory
     case settings
 }